feat(user-api,shopping-ai,product-api): finalizei o Capítulo 9 do projeto backend mapeando o status code de cada rota de forma adequada e adicionando o tratamento de exceptions e também adicionei uma validação inicial para cpf por meio do banco de dados.

+++ services/shopping-api/src/main/java/com/microsservicos/shoppingapi/repository/impl/ReportRepositoryImpl.java
@@ -1,6 +1,5 @@
 package com.microsservicos.shoppingapi.repository.impl;
 
-import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -60,7 +59,7 @@ public class ReportRepositoryImpl implements ReportRepository {
 
     Object[] result = (Object[]) query.getSingleResult();
 
-    return new ReportDto((BigInteger) result[0], (Double) result[1],  (Double) result[2]);
+    return new ReportDto((Long) result[0], (Double) result[1],  (Double) result[2]);
   }
 
 }
