+++ services/utils/src/main/java/com/microsservicos/dto/ReportDto.java
@@ -1,5 +1,3 @@
 package com.microsservicos.dto;
 
-import java.math.BigInteger;
-
-public record ReportDto(BigInteger count, Double total, Double mean) {}
+public record ReportDto(Long count, Double total, Double mean) {}
